new simplified resource model

+++ src/yae/string.cpp
@@ -8,8 +8,8 @@ namespace string {
 
 size_t safeCopyToBuffer(char* _destination, const char* _source, size_t _destinationSize)
 {
-	size_t sourceLength = strlen(_source);
-	size_t copiedLength = math::min(sourceLength, _destinationSize - 1);
+	u32 sourceLength = strlen(_source);
+	u32 copiedLength = math::min(sourceLength, u32(_destinationSize) - 1);
 	strncpy(_destination, _source, copiedLength);
 	_destination[copiedLength] = 0;
 	return copiedLength;
